+++ app/Http/Controllers/API/UserVideoController.php
@@ -1,8 +1,11 @@
 <?php
+
 namespace App\Http\Controllers\API;
+
 use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
 use App\UserVideo;
+
 class UserVideoController extends BaseController
 {
     public function __construct()
@@ -11,14 +14,18 @@ class UserVideoController extends BaseController
     }
     public function index()
     {
-        $items = UserVideo::all();
-        return $this->sendResponse($items->toArray(), 'Albums retrieved successfully.');
+        $items = UserVideo::where('user_videos.user_id', auth()->user()->id)
+            ->join('videos', 'videos.id', '=', 'user_videos.video_id')
+            ->join('artists', 'videos.artist_id', '=', 'artists.id')
+            ->select('videos.*', 'artists.name as artistName')
+            ->get();
+        return $this->sendResponse($items, 'Albums retrieved successfully.');
     }
     public function store(Request $request)
     {
         $input = $request->all();
 
-        $item = UserVideo::create(["user_id"=>$request->user()->id,"video_id"=>$input["id"]]);
+        $item = UserVideo::create(["user_id" => $request->user()->id, "video_id" => $input["id"]]);
         return $this->sendResponse($item->toArray(), 'Album created successfully.');
     }
     public function show($id)
@@ -29,16 +36,16 @@ class UserVideoController extends BaseController
         }
         return $this->sendResponse($item->toArray(), 'Album retrieved successfully.');
     }
-    public function update(Request $request, UserVideo $item)
+    // public function update(Request $request, UserVideo $item)
+    // {
+    //     $input = $request->all();
+    //     $item->name = $input['name'];
+    //     $item->save();
+    //     return $this->sendResponse($item->toArray(), 'Product updated successfully.');
+    // }
+    public function destroy($id)
     {
-        $input = $request->all();
-        $item->name = $input['name'];
-        $item->save();
-        return $this->sendResponse($item->toArray(), 'Product updated successfully.');
-    }
-    public function destroy(UserVideo $item)
-    {
-        $item->delete();
-        return $this->sendResponse($item->toArray(), 'Album deleted successfully.');
+        $item = UserVideo::where(["user_id" => auth()->user()->id, "video_id" => $id])->delete();
+        return $this->sendResponse($item, 'Album deleted successfully.');
     }
 }
